refactor fullsizePhoto.js, fix errors and bugs
