+++ js/data.js
@@ -2,7 +2,7 @@ export const MIN_DESCRIPTION_IDENTIFIER = 1;
 export const MAX_DESCRIPTION_IDENTIFIER = 25;
 
 export const MIN_NUMBER_OF_LIKES = 15;
-export const MAX_NUMBER_OF_LIKES = 200;
+export const MAX_NUMBER_OF_LIKES = 25;
 
 export const MIN_NUMBER_FOR_LIKES = 1;
 export const MAX_NUMBER_FOR_LIKES = 6;
@@ -20,6 +20,8 @@ export const REDUCE_SCALE_VALUE = '0.25';
 export const VALUE_OF_DECIMAL_NUMBER_SYSTEM = 10;
 
 export const REGULAR_EXPRESSION_FOR_HASHTAGS = /^#[A-Za-zА-Яа-яЁё0-9]{1,19}$/;
+
+export const MAX_ADD_COMMENTS = 5;
 
 export const DESCRIPTIONS = [
   'Весело',
