+++ js/main.js
@@ -1,5 +1,5 @@
 import './miniatureDrawing.js';
-import './fullsizePhotoOpen.js';
+import './fullsizePhoto.js';
 import './form.js';
 import './scale.js';
 import './slider.js';
